requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
numpy==1.26.0
opencv-python==4.8.1.78
Pillow==10.0.1
requests==2.31.0
torch==1.11.0+cu113
torchvision==0.12.0
typing_extensions==4.8.0
urllib3==2.0.6
